gui/requirements.txt
arabic_reshaper==2.1.0
argh==0.26.2
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
# cryptography==1.7.1
Cython==0.29.10
dbus-python==1.2.16
decorator==4.4.2
docutils==0.16
evento==1.0.2
future==0.18.2
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
kafka-python==2.0.1
keyring==10.1
keyrings.alt==1.3
# Kivy==1.10.1
# Kivy-Garden==0.1.4
oauth2client==4.1.3
oauthlib==3.1.0
omxplayer-wrapper==0.3.3
pathlib2==2.3.5
Pillow==7.2.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
# pycrypto==2.6.1
# pygame==1.9.6
Pygments==2.6.1
# pygobject==3.22.0
pyserial==3.4
python-bidi==0.4.2
pytz==2020.1
# pyxdg==0.25
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
SecretStorage==2.3.1
six==1.10.0
tweepy==3.9.0
uritemplate==3.0.1
urllib3==1.25.10
watchdog==0.10.3
